+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- [#171](https://github.com/deviceinsight/kafkactl/issues/171) Support for reset command to reset offset to offset from provided datetime
+- [#172](https://github.com/deviceinsight/kafkactl/issues/172) Support for json format in produce command
+
 ## 3.4.0 - 2023-09-25
 
 ### Fixed